web-app/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio
fastapi
uvicorn
python-multipart
numpy
opencv-python
pillow
python-jose
pydantic 
ultralytics
albumentations
scikit-learn